push all python checks in language.target_info
